# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Nexus evidence workbench",
    "version": "0.2.0"
  },
  "paths": {
    "/api/investigation": {
      "get": {
        "summary": "Get Investigation",
        "operationId": "get_investigation_api_investigation_get",
        "parameters": [
          {
            "name": "case_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "kamoa-to-cables",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Investigation"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/entities/{entity_id}/neighborhood": {
      "get": {
        "summary": "Get Neighborhood",
        "operationId": "get_neighborhood_api_entities__entity_id__neighborhood_get",
        "parameters": [
          {
            "name": "entity_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Entity Id"
            }
          },
          {
            "name": "depth",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 2,
              "minimum": 1,
              "default": 1,
              "title": "Depth"
            }
          },
          {
            "name": "case_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "kamoa-to-cables",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Neighborhood"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/claims/{claim_id}/projection": {
      "get": {
        "summary": "Claim Projection",
        "operationId": "claim_projection_api_claims__claim_id__projection_get",
        "parameters": [
          {
            "name": "claim_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Claim Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ClaimProjection"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Claim Projection Api Claims  Claim Id  Projection Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/entities/resolve": {
      "get": {
        "summary": "Resolve Entity",
        "operationId": "resolve_entity_api_entities_resolve_get",
        "parameters": [
          {
            "name": "case_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          },
          {
            "name": "legacy_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Legacy Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/EntityRecord"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Resolve Entity Api Entities Resolve Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/measurements": {
      "get": {
        "summary": "Measurements",
        "operationId": "measurements_api_measurements_get",
        "parameters": [
          {
            "name": "entity",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Entity"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MeasurementRecord"
                  },
                  "title": "Response Measurements Api Measurements Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/claims/{claim_id}/history": {
      "get": {
        "summary": "History",
        "operationId": "history_api_claims__claim_id__history_get",
        "parameters": [
          {
            "name": "claim_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Claim Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewHistory"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/claims/{claim_id}/proposals": {
      "post": {
        "summary": "Propose",
        "operationId": "propose_api_claims__claim_id__proposals_post",
        "parameters": [
          {
            "name": "claim_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Claim Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProposalRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewHistory"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/proposals/{proposal_id}/decision": {
      "post": {
        "summary": "Decide",
        "operationId": "decide_api_proposals__proposal_id__decision_post",
        "parameters": [
          {
            "name": "proposal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Proposal Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewHistory"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Alias": {
        "properties": {
          "case_id": {
            "type": "string",
            "title": "Case Id"
          },
          "legacy_id": {
            "type": "string",
            "title": "Legacy Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "case_id",
          "legacy_id"
        ],
        "title": "Alias"
      },
...
      "ClaimProjection": {
        "properties": {
          "release_id": {
            "type": "string",
            "title": "Release Id"
          },
          "basis_proposal_id": {
            "type": "string",
            "format": "uuid",
            "title": "Basis Proposal Id"
          },
          "claim": {
            "$ref": "#/components/schemas/ClaimRecord"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "release_id",
          "basis_proposal_id",
          "claim"
        ],
        "title": "ClaimProjection"
      },
...
      "DecisionRequest": {
        "properties": {
          "request_id": {
            "type": "string",
            "format": "uuid",
            "title": "Request Id"
          },
          "expected_revision": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Expected Revision"
          },
          "decision": {
            "type": "string",
            "enum": [
              "accept",
              "reject"
            ],
            "title": "Decision"
          },
          "reviewer": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Reviewer"
          },
          "reason": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 1,
            "title": "Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "request_id",
          "expected_revision",
          "decision",
          "reviewer",
          "reason"
        ],
        "title": "DecisionRequest"
      },
      "Entity": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "coordinates": {
            "anyOf": [
              {
                "prefixItems": [
                  {
                    "type": "number"
                  },
                  {
                    "type": "number"
                  }
                ],
                "type": "array",
                "maxItems": 2,
                "minItems": 2
              },
              {
                "type": "null"
              }
            ],
            "title": "Coordinates"
          },
          "location_note": {
            "type": "string",
            "title": "Location Note"
          },
          "location_precision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Precision"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "name",
          "type",
          "coordinates",
          "location_note"
        ],
        "title": "Entity"
      },
      "EntityRecord": {
        "properties": {
          "canonical_id": {
            "type": "string",
            "title": "Canonical Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "subtype": {
            "type": "string",
            "title": "Subtype"
          },
          "aliases": {
            "items": {
              "$ref": "#/components/schemas/Alias"
            },
            "type": "array",
            "title": "Aliases"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "canonical_id",
          "name",
          "category",
          "subtype",
          "aliases"
        ],
        "title": "EntityRecord"
      },
...
      "EvidencePack": {
        "properties": {
          "schema_version": {
            "type": "string",
            "title": "Schema Version"
          },
          "case_id": {
            "type": "string",
            "title": "Case Id"
          },
          "checked_on": {
            "type": "string",
            "format": "date",
            "title": "Checked On"
          },
          "scope": {
            "type": "string",
            "title": "Scope"
          },
          "verdict": {
            "type": "string",
            "title": "Verdict"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/Source"
            },
            "type": "array",
            "title": "Sources"
          },
          "entities": {
            "items": {
              "$ref": "#/components/schemas/Entity"
            },
            "type": "array",
            "title": "Entities"
          },
          "claims": {
            "items": {
              "$ref": "#/components/schemas/Claim"
            },
            "type": "array",
            "title": "Claims"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/Event"
            },
            "type": "array",
            "title": "Events"
          },
          "excluded_assertions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Excluded Assertions",
            "default": []
          },
          "briefing": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Briefing"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "schema_version",
          "case_id",
          "checked_on",
          "scope",
          "verdict",
          "sources",
          "entities",
          "claims",
          "events"
        ],
        "title": "EvidencePack"
      },
...
      "GuidedGeography": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "framing": {
            "type": "string",
            "title": "Framing"
          },
          "stops": {
            "items": {
              "$ref": "#/components/schemas/GeoStop"
            },
            "type": "array",
            "title": "Stops"
          },
          "routes": {
            "items": {
              "$ref": "#/components/schemas/GeoRoute"
            },
            "type": "array",
            "title": "Routes",
            "default": []
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title",
          "framing",
          "stops"
        ],
        "title": "GuidedGeography"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Investigation": {
        "properties": {
          "pack": {
            "$ref": "#/components/schemas/EvidencePack"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/ReadingPrompt"
            },
            "type": "array",
            "title": "Questions"
          },
          "geography": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GuidedGeography"
              },
              {
                "type": "null"
              }
            ]
          },
          "answer_mode": {
            "type": "string",
            "title": "Answer Mode",
            "default": "curated_reading_guide"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "pack",
          "questions"
        ],
        "title": "Investigation"
      },
      "MeasurementRecord": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "case_id": {
            "type": "string",
            "title": "Case Id"
          },
          "metric_id": {
            "type": "string",
            "title": "Metric Id"
          },
          "point_index": {
            "type": "integer",
            "title": "Point Index"
          },
          "measure": {
            "type": "string",
            "title": "Measure"
          },
          "value_text": {
            "type": "string",
            "title": "Value Text"
          },
          "entity": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entity"
          },
          "original_label": {
            "type": "string",
            "title": "Original Label"
          },
          "temporal": {
            "$ref": "#/components/schemas/Temporal"
          },
          "epistemic_status": {
            "type": "string",
            "title": "Epistemic Status"
          },
          "release_status": {
            "type": "string",
            "title": "Release Status"
          },
          "claim_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Claim Ids"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "case_id",
          "metric_id",
          "point_index",
          "measure",
          "value_text",
          "entity",
          "original_label",
          "temporal",
          "epistemic_status",
          "release_status",
          "claim_ids",
          "note"
        ],
        "title": "MeasurementRecord"
      },
...
      "Neighborhood": {
        "properties": {
          "entities": {
            "items": {
              "$ref": "#/components/schemas/Entity"
            },
            "type": "array",
            "title": "Entities"
          },
          "claims": {
            "items": {
              "$ref": "#/components/schemas/Claim"
            },
            "type": "array",
            "title": "Claims"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/Source"
            },
            "type": "array",
            "title": "Sources"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "entities",
          "claims",
          "sources"
        ],
        "title": "Neighborhood"
      },
...
      "ProposalRequest": {
        "properties": {
          "request_id": {
            "type": "string",
            "format": "uuid",
            "title": "Request Id"
          },
          "base_revision": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Base Revision"
          },
          "statement": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 1,
            "title": "Statement"
          },
          "caveat": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 1,
            "title": "Caveat"
          },
          "evidence_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 30,
            "minItems": 1,
            "title": "Evidence Ids"
          },
          "valid_from": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid From"
          },
          "valid_to": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid To"
          },
          "author": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Author"
          },
          "reason": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 1,
            "title": "Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "request_id",
          "base_revision",
          "statement",
          "caveat",
          "evidence_ids",
          "author",
          "reason"
        ],
        "title": "ProposalRequest"
      },
      "ReadingPrompt": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "question": {
            "type": "string",
            "title": "Question"
          },
          "expected_answer": {
            "type": "string",
            "title": "Expected Answer"
          },
          "claim_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Claim Ids"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "question",
          "expected_answer",
          "claim_ids",
          "status"
        ],
        "title": "ReadingPrompt"
      },
...
      "ReviewHistory": {
        "properties": {
          "claim_id": {
            "type": "string",
            "title": "Claim Id"
          },
          "current_revision": {
            "type": "integer",
            "title": "Current Revision"
          },
          "proposals": {
            "items": {
              "$ref": "#/components/schemas/Proposal"
            },
            "type": "array",
            "title": "Proposals"
          },
          "versions": {
            "items": {
              "$ref": "#/components/schemas/AcceptedVersion"
            },
            "type": "array",
            "title": "Versions"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "claim_id",
          "current_revision",
          "proposals",
          "versions"
        ],
        "title": "ReviewHistory"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}